\name{pathvisioServer}
\alias{pathvisioServer}
\title{
pathvisio Server function
}
\description{
Function for hosting a server which enables the link between R and the pathvisioRpc Jar file
}
\usage{
pathvisioServer(port=9000)
}
\arguments{
 \item{port}{communication port on which PathVisioRPC can be contacted}
}
\author{
Anwesha Dutta
Egon Willighagen
Christ Leemans

Maintainer: Anwesha Dutta <anwesha.dutta@maastrichtuniversity.nl>
<Egon Willighagen <egon.willighagen@gmail.com>
<C.Leemans@student.maastrichtuniversity.nl>
}