\name{pvxmlrpc-package}
\alias{pvxmlrpc-package}
\alias{pvxmlrpc}
\docType{package}
\title{
R package for PathVisio XML-RPC
}
\description{
Interactively create, edit, and visualize pathways with PathVisio.
}
\details{
This package wraps the PathVisio XML-RPC code with a R specific API.
}
\author{
Egon Willighagen

Maintainer: Egon Willighagen <egon.willighagen@gmail.com>
}
\keyword{ package }
\keyword{ PathVisio }
\keyword{ XML-RPC }
