\name{RPathVisio-package}
\alias{RPathVisio-package}
\alias{RPathVisio}
\docType{package}
\title{
R package for PathVisio XML-RPC
}
\description{ 
Directly call functions of PathVisio to create and edit biological pathways, visualize data on pathways, perform pathway statistics and get a list of participating genes/metabolites/reactions etc. in a pathway.
\details{
This package wraps the PathVisio XML-RPC code with a R specific API.
}
\author{
Anwesha Dutta
Egon Willighagen
Christ Leemans

Maintainer: Anwesha Dutta <anwesha.dutta@maastrichtuniversity.nl>
<Egon Willighagen <egon.willighagen@gmail.com>
<C.Leemans@student.maastrichtuniversity.nl>
}
\keyword{ package }
\keyword{ PathVisio }
\keyword{ XML-RPC }
